Added sorting method and click handler so that the name column and type column would sort in alphabetical order and reverse when their respected header is clicked
 Created variable sorted that  saves a boolean. The boolean tells if the column has been sorted or not.

In sorted method, this.sorted = !this.sorted;is added to flip the sorted value true or false. This allows the column to flip back and forth each time header is click rather than just once.

a second if else statement added where the first if/else   returns if sorted = true and the second one returns if sorted = false.